requirements.txt
chinese_calendar
lxml
openai>1.0.0
pandas
pyyaml
requests
schedule
pyhandytools
sparkdesk-api==1.3.0
wcferry>=39.0.10.0
websocket
pillow
jupyter_client
zhdate
ipykernel
google-generativeai
zhipuai
faker
bs4
selenium